update login test, move mock files, add mock fetch to own file
