+++ src/js/mocks/fetch.mock.js
@@ -1,0 +1,9 @@
+export const createMockFetch = (response = {}, status = 200) => {
+  return jest.fn().mockImplementation(() =>
+    Promise.resolve({
+      ok: status < 400,
+      status,
+      json: () => Promise.resolve(response),
+    }),
+  );
+};
