+++ src/js/mocks/localStorage.mock.js
@@ -4,8 +4,8 @@ export const localStorageMock = () => {
   return {
     getItem: jest.fn((key) => storage[key] || null),
     setItem: jest.fn((key, value) => {
-      (storage[key] = value.toString()), console.log("key", key);
-      console.log("value", value);
+      (storage[key] = value.toString()), console.log('key', key);
+      console.log('value', value);
     }),
     removeItem: jest.fn((key) => {
       delete storage[key];
