+++ src/js/mocks/user.mock.js
@@ -1,7 +1,7 @@
 const name = 'Test User';
 const email = 'test@stud.noroff.no';
 const password = 'passwordTest123';
-export const accessToken = 'tokenTest';
+const accessToken = 'tokenTest';
 
 export const userData = {
   name: name,
